requirements.txt
googletrans==4.0.0rc1
gTTS==2.5.1
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.31.1
transformers==4.38.2